+++ ChangeLog.md
@@ -8,6 +8,7 @@
 ## [0.2] (future release)
 * accept self-signed certificates
 * code validated with shellcheck
+* switch API endpoint from /store to /envelope
 
 ## [0.1] - 2025-02-22
 * function documentation